Refactor mermaid renderer structure
