+++ src/marp.ts
@@ -10,7 +10,7 @@ import * as emojiPlugin from './emoji/emoji'
 import { defaultHTMLAllowList, type HTMLAllowList } from './html/allowlist'
 import * as htmlPlugin from './html/html'
 import * as mathPlugin from './math/math'
-import * as mermaid from './mermaid'
+import * as mermaid from './mermaid/mermaid'
 import * as scriptPlugin from './script/script'
 import * as shiki from './shiki'
 import * as sizePlugin from './size/size'
